Relay token / error handles to exclusive functions
